Update bash script for DB dump

+++ travis/dump.sh
@@ -1,5 +1,11 @@
-mysqldump -uroot -p --add-drop-database --disable-keys --databases directorzone_zf2 > /home/cmoreton/db_create.sql
-cd /home/cmoreton
-rm db_create.zip
-zip db_create.zip db_create.sql
-rm db_create.sql
+#!/bin/bash
+
+if [ `hostname -s` = "vagrant" ]; then 
+  exit 0
+else
+  mysqldump -uroot -p --add-drop-database --disable-keys --databases directorzone_zf2 > /home/cmoreton/db_create.sql
+  cd /home/cmoreton
+  rm db_create.zip
+  zip db_create.zip db_create.sql
+  rm db_create.sql
+fi
